Add files via upload
In this project, I learned the basics of Tableau. I went to Kaggle and downloaded a data set on Airbnbs in Seattle. The use case would be my uncle wants to move into a new Airbnb market. So he chose Seatle. He wanted me to show what was the best area to have houses in and how many bedrooms he would need to maximize profit. Since there's is a lot of data in the data set I downloaded I had to drop the reviews table.  The reviews table really wouldn't matter to my uncle because they are reviews from other people's rentals and he's been in the business for years so he kinda an idea of what a renter wants and needs. The 2 remaining tables are calendars and listings. The first sheet is made of a color-coded bar graph comparing the average price per zip code. The next sheet consists of a fancy map with a color-coded area representing the zipcodes. In the file, you can hover over the color to see the average price of Airbnb rental in that area. The next sheet shows revenue per year. I added the sum price and date from the calendar table to make the line graph. It makes sense that in the summer months the revenue is higher because a lot of people travel during the summer. In sheet number 4 I bar a bar graph comparing the average cost per bedroom. And I guessed the more the bedrooms the higher the rental price. In the next sheet, I just did a normal table with the distinct number of rentals with how many bedrooms each have. For example, there are only 5 rentals with 6 bedrooms.  this is something I wanted to add just to put it don't the dashboard so when my Uncle sees it he can see how many of each rental type there are. And finally, on the dashboard, I have put together all of the previous graphs and charts int to one. If my Uncle wants to see what area the average price per zip code or where the zip code is located on the map he can simply hover his mouse over the respective shaded area on the map or the bar on the graph.