Add files via upload
This is a group project I did in my Database Systems class with my 2 classmates, Scott Kinnie and Dominic Krajicek. In this project, we were tasked with creating a database for any company. We decided on a clothing company so we chose Yakuzaboss.com. They are an anime-inspired clothing company. In the project, we made a business description and Entities Relationship diagram, a database Design diagram as well as SQL Create Table and Insert statements.